Add argument order foot-gun avoidance constructors (#8)
The Internet pointed out that it's easy to get the argument order mixed
up for constructors like `Bearing::new` (does azimuth or elevation come
first) and `Wgs84::new` (does latitude or longitude come first). To that
end, this PR adds two new construction mechanisms for both of these,
while also deprecating (but not removing) `::new` in favor of those. The
two new constructors are:

- [Builders] that make fields be set by name (eg, `.azimuth(az)`) while
  statically enforcing (using the type-state pattern) that all fields
  are set by the time the final object is constructed.
- Single-struct-operand constructors that just take a new struct type
  whose fields correspond to the arguments (eg, `Components { azimuth:
  Angle, elevation: Angle }`).

With both of these, callers must explicitly name each parameter, and
thus cannot get their order wrong.

The `from_cartesian` constructor for `Coordinate` and `Vector` both take
three arguments of the exact same type, and thus have a similar
challenge. While it's less likely that someone passes NED coordinates in
anything but N E D order, in the interest of foot-gun avoidance it felt
right to give those constructors a similar make-over. So, `Coordinate`
and `Vector` both gained the same two constructors. They also both
gained a macro constructor that takes single-character-indicated
components (eg, `coordinate!(n = 1, e = 0, d = 5)`) and fails if the
indicated components do not match the target coordinate system
convention (eg, providing `n =` to an `FrdLike` system).

Note that this PR does _not_ add similar safety construction to the
`from_spherical` or `from_tait_bryan_angles` constructors. For the
former, doing so _might_ make sense, but it would be fine to add that in
a follow-up PR. For the latter, splitting out the named arguments may be
counter-productive; the rotations are _intrinsic_, so if they're
specified in a different order (eg, making it a `struct` such that
`roll` could be specified first) might give the reader the wrong
impression.

This PR _also_ does not type every component (eg, making a newtype for
`Azimuth`). This is in part because it would make the interface a whole
lot noisier and less ergonomic (which can itself cause bugs as the
calling code becomes less readable), and in part because it would add
only limited safety as it would still be possible to provide the wrong
`Angle` to `Azimuth::new`.

To avoid making this a backwards-incompatible change, as mentioned this
PR retains the `from_cartesian` and `new` constructors. It marks them as
deprecated to encourage users to upgrade without forcing them to do so.
In the next major release, those constructors will likely be removed
from the public API.

[Builders]: https://rust-unofficial.github.io/patterns/patterns/creational/builder.html