feat: add Vector::orientation_at_origin
Without this, it was very inconvenient to go from XYZ coordinates to the
orientation of an object pointed in the direction of that vector. Some
users would end up using `bearing_at_origin`, but then get bitten by
bearing azimuth and elevation not matching yaw-pitch as expected by the
Tait-Bryan constructors for ENU coordinate systems.

Also fixes #36 as a drive-by.